requirements.txt
# Also Requires gobject-introspection, Gtk 4, and LibAdwaita 1
pyyaml
sigfig
logging42